Move new exceptions
