+++ src/Exception/Api/ForbiddenException.php
@@ -14,20 +14,17 @@ declare(strict_types=1);
  *  @license    http://www.pimcore.org/license     GPLv3 and PCL
  */
 
-namespace Pimcore\Bundle\StudioBackendBundle\Exception;
+namespace Pimcore\Bundle\StudioBackendBundle\Exception\Api;
 
 use Pimcore\Bundle\StudioBackendBundle\Util\Constants\HttpResponseCodes;
 
 /**
  * @internal
  */
-final class InvalidThumbnailException extends AbstractApiException
+final class ForbiddenException extends AbstractApiException
 {
-    public function __construct(string $thumbnailName)
+    public function __construct(string $message = 'Access Denied')
     {
-        parent::__construct(
-            HttpResponseCodes::BAD_REQUEST->value,
-            sprintf('Invalid thumbnail: %s', $thumbnailName)
-        );
+        parent::__construct(HttpResponseCodes::FORBIDDEN->value, $message);
     }
 }
