+++ src/Workflow/Service/WorkflowGraphService.php
@@ -16,9 +16,7 @@ declare(strict_types=1);
 
 namespace Pimcore\Bundle\StudioBackendBundle\Workflow\Service;
 
-use Exception;
 use Pimcore\Bundle\StudioBackendBundle\Exception\Api\WorkflowDependencyMissingException;
-use Pimcore\Bundle\StudioBackendBundle\Exception\ConsoleDependencyMissingException;
 use Pimcore\Bundle\StudioBackendBundle\Util\Traits\ConsoleExecutableTrait;
 use Pimcore\Model\Element\ElementInterface;
 use Pimcore\Tool\Console;
